Set a default PATH for pass execution, fixes #2
This uses macOS' path_helper tool to identify the default PATH for the
system. If the user doesn't provide a PATH env var in the config file
then the system default PATH is provided as an env var for pass'
execution.

This should fix bugs where the pass, gpg, tree and getopt binaries
aren't found by default.